feat: major Helioponic system overhaul - smart automation, animated UI, and redesigned monitoring
## Backend
- Refactor sensor model: pompa1/pompa2 become Optional (None = simulator path)
- Add pH threshold support (ph_min/ph_max) for Pompa 2 pH dosing automation
- Implement manual override system with 30s cooldown in analytics router
- Add hysteresis state tracking and reset in MQTT subscriber
- Wire callbacks for manual override, hysteresis reset, and notification state reset
- Add unread notification count endpoint
- Update automation engine to support pH-based pump control
- Remove legacy energy model/service (replaced by aggregated energy tracking)

## Firmware
- Update ESP32 example firmware with pH sensor support and new threshold variables

## Mobile App
- Redesign WaterWaveWidget: replace circular SVG with rectangular plastic tank side view
- Add smooth animated water level transitions (500ms Animated.timing)
- Create reusable AnimatedProgressBar component for smooth progress bar animations
- Apply animated progress bars to DashboardScreen (4 bars) and PIDScreen (3 bars)
- Redesign WaterCard with gradient accent bar, status badges, stats grid
- Redesign Sensor Dashboard in PIDScreen with 3-column grid, progress bars, status badges
- Redesign Recent Activity section with timeline-style feed, mini bars, pump badges
- Add smart rule activation: toggling a rule auto-enables + isolates that rule
- Add greyscale disabled UI for Automation cards when auto-pump is OFF
- Add automation callbacks: manual override disables auto, immediate state reflection
- Add formatRelativeTime helper for cleaner timestamps in activity feed
- Update API client with notifications, automation, and night mode endpoints
- Update sensor store with override pump states and auto-enabled state
- Add night mode store and notification store for real-time updates
- Clean up unused EnergyDonutChart component and energyStore

## Tools
- Update simulation scripts with pH sensor data, manual override testing
- Update seed data with pH thresholds
- Improve data diagnosis and verification tools

## Chore
- Add nul to .gitignore (Windows system device file)

+++ helioponic-backend/app/models/water.py
@@ -25,7 +25,7 @@ class WaterRecord(BaseModel):
     """
     device_id: str = "HELIO_001"
     recorded_at: datetime = Field(default_factory=lambda: datetime.now(UTC))
-    jarak_cm: int = 999
+    jarak_cm: float = 999.0
     water_level_pct: float = 0.0
 
     class Settings:
